밴유저 대화기록 저장이슈 해결

+++ lib/T_card.dart
@@ -322,15 +322,49 @@ class _TCardPageState extends State<TCardPage> {
                                   TextButton(
                                     child: const Text('확인',style: TextStyle(fontFamily: "Galmuri11-Bold",color: Colors.deepOrange),),
                                     onPressed: () async{
-                                      var collection = FirebaseFirestore.instance.collection(widget.fingerPrint!+widget.fingerprintkeys[_index]);
-                                      var snapshots = await collection.get();
-                                      for (var doc in snapshots.docs) {
-                                        await doc.reference.delete();
+                                      bool bannedUser=false;
+                                      var collection_ban = FirebaseFirestore.instance.collection("bannedUser");
+                                      var snapshots_ban = await collection_ban.get();
+                                      for (var doc in snapshots_ban.docs) {
+                                        var banneduser= await doc.reference.get();
+                                        if(widget.fingerPrint==banneduser["fingerPrint"])
+                                        {
+                                          bannedUser=true;
+                                          print("true");
+                                        }
+                                      }
+                                      if(bannedUser==false)
+                                      {
+                                        var collection = FirebaseFirestore.instance.collection(widget.fingerPrint!+widget.fingerprintkeys[_index]);
+                                        var snapshots = await collection.get();
+                                        for (var doc in snapshots.docs) {
+                                          await doc.reference.delete();
+                                        }
+                                      }
+                                      else{
+                                        showDialog(
+                                            context: context,
+                                            barrierDismissible: true,
+                                            builder: (BuildContext context) {
+                                              return AlertDialog(
+                                                content: Text("당신은 밴유저입니다.\n\n2019037038로 문의주십시오.",style: TextStyle(fontFamily:"Galmuri11-Bold",fontSize: 20,color: Colors.red),),
+                                                insetPadding: const  EdgeInsets.fromLTRB(0,80,0, 80),
+                                                actions: [
+                                                  TextButton(
+                                                    child: const Text('확인',style: TextStyle(fontFamily: "Galmuri11-Bold",color: Colors.red),),
+                                                    onPressed: () async{
+                                                      Navigator.of(context).pop();
+                                                    },
+                                                  ),
+                                                ],
+                                              );
+                                            }
+                                        );
                                       }
                                       deleteData();
                                       Navigator.of(context).pop();
                                       Navigator.of(context).pop();
-                                    },
+                                    }
                                   ),
                                 ],
                               );
